Refactor common code into a RowHeaderCell class
