+++ analytics_dashboard/static/apps/course-list/list/views/course-id-and-name-cell.js
@@ -6,23 +6,17 @@ define(function(require) {
     'use strict';
 
     var _ = require('underscore'),
-        Backgrid = require('backgrid'),
 
+        RowHeaderCell = require('components/generic-list/list/views/row-header-cell'),
         courseIdAndNameCellTemplate = require('text!course-list/list/templates/course-id-and-name-cell.underscore'),
 
         CourseIdAndNameCell;
 
-    CourseIdAndNameCell = Backgrid.Cell.extend({
-        tagName: 'th',
+    CourseIdAndNameCell = RowHeaderCell.extend({
         className: 'course-name-cell',
         template: _.template(courseIdAndNameCellTemplate),
         events: {
             click: 'emitTracking'
-        },
-        render: function() {
-            this.$el.html(this.template(this.model.toJSON()));
-            this.$el.attr('scope', 'row');
-            return this;
         },
         emitTracking: function() {
             this.$el.find('a').trigger('clicked.tracking');
